some doc & reorganize the Interface
