+++ src/transformers/masking_utils.py
@@ -20,7 +20,7 @@ import torch.nn.functional as F
 
 from .cache_utils import Cache
 from .configuration_utils import PretrainedConfig
-from .modeling_utils import GeneralInterface
+from .utils.generic import GeneralInterface
 from .utils.import_utils import is_torch_flex_attn_available, is_torch_greater_or_equal, is_torchdynamo_compiling
